+++ NEWS.md
@@ -21,6 +21,7 @@
   - Fix Notes parsing for Comictagger beta Metron origin
   - ComicInfo.xml orders tags properly per the xsd.
   - Corrected schemaLocation tags for xml formats to be valid.
+  - ComicBookInfo.json `rating`, and `tags` tags fixed.
 
 ## v1.2.1